requirements.txt
mccabe==0.6.1
pycodestyle==2.7.0
pydocstyle==5.1.1
pyflakes==2.3.0
pylama==7.7.1
snowballstemmer==2.1.0
psycopg2-binary
